refactor(cli): split memory.rs and main.rs so Phases 2 and 3 can run in parallel
Pure structural move, no behavior change. The adaptive-context plan sanctions
this explicitly (docs/design/adaptive-context-plan.md:33-35): a phase that must
grow a file already at its ratchet baseline splits that file first, as a pure
move.

Phases 2 (#713) and 3 (#714) are being built on parallel branches. Their
deliverables are disjoint, but three shared files would have collided in ways a
textual merge resolves *wrongly*:

- `stella-cli/src/memory.rs` (929 lines) held both the recall plane (Phase 2's
  telemetry and provenance work) and the learning loop (Phase 3's observation
  and mining work). Two branches rewriting opposite halves of one file meet in
  `SessionMemory` and merge clean while being wrong. Split along that seam into
  `memory/recall.rs` and `memory/learning.rs`; `memory.rs` keeps the type, the
  constructors, and the episode/taxonomy recording.

- `stella-cli/src/main.rs` sat at exactly its 1796-line ceiling with zero
  headroom, and both phases add subcommand surface. Moved storage, scripts, and
  graph commands into owning modules, following the pattern already used by
  stats.rs and usage_cmd.rs. It now lands at 1337 — under the 1500 limit, so its
  baseline entry is deleted rather than raised.

- Both migration ladders are silent-merge hazards: store.db's is an index-ordered
  array where SCHEMA_VERSION = MIGRATIONS.len(), and context.db's is an
  `if version < N` chain. Two branches each appending "the next slot" produce a
  broken ladder that git merges without complaint. Slots are now reserved by
  comment at both append points.

Behavior neutrality was checked two ways: every removed line reappears (the only
exceptions are visibility widenings to `pub(super)`, path qualifications, and the
deleted baseline row), and the generated clap help for every moved command is
byte-identical across 387 lines.

`make gate` green: 61 test suites, file-size OK at 27 grandfathered files.

Refs #713, #714, #469
