+++ stella-cli/src/memory.rs
@@ -29,28 +29,40 @@ use std::path::{Path, PathBuf};
 use colored::Colorize;
 use serde::{Deserialize, Serialize};
 use stella_context::{
-    ContextDelta, ContextQuery, ContextStore, DomainInput, EpisodeInput, EpisodeOutcome,
-    FactAssertion, HashEmbedder, MemoryInput, NodeInput, NodeKind, SystemClock, format_rfc3339,
+    ContextDelta, ContextStore, DomainInput, EpisodeInput, EpisodeOutcome, FactAssertion,
+    HashEmbedder, NodeInput, NodeKind, SystemClock, format_rfc3339,
 };
-use stella_core::skills::{
-    self, AutoCreateConfig, AutoCreateDecision, SelectionConfig, Skill, SkillMineConfig,
-    SkillObservation,
-};
-use stella_pipeline::{ContextRecallPort, Recall, RecalledFrame};
-use stella_protocol::{CompletionMessage, MessageRole, Provider};
+use stella_core::skills::{self, SelectionConfig, Skill};
 
 use crate::domains::Domains;
 
+// Reached only through `use super::*` in the sibling test modules
+// (`memory/tests.rs`, `memory/quarantine_tests.rs`): the code that used
+// these moved to `memory/recall.rs` and `memory/learning.rs`, the tests
+// deliberately did not.
+#[cfg(test)]
+use stella_context::MemoryInput;
+#[cfg(test)]
+use stella_pipeline::{ContextRecallPort, RecalledFrame};
+#[cfg(test)]
+use stella_protocol::{CompletionMessage, MessageRole};
+
+mod learning;
 mod private_state;
 mod projection;
 #[cfg(test)]
 mod quarantine_tests;
+mod recall;
 #[path = "memory/skills.rs"]
 mod skill_files;
 mod suppression;
 mod tuning;
 use private_state::resolve_context_db_path;
+#[cfg(test)]
 use projection::{is_suppressed_local_frame, project_recalled_frame};
+pub use recall::inject_recall_block;
+#[cfg(test)]
+use recall::{ab_control_turn, goal_path_anchors, render_context_section};
 #[cfg(test)]
 pub(crate) use skill_files::load_workspace_skills;
 pub(crate) use skill_files::{load_workspace_skills_with_authority, workspace_skills_dir};
@@ -237,98 +249,6 @@ impl SessionMemory {
         load_workspace_skills_with_authority(&self.workspace_root, self.include_workspace_skills)
             .skills
     }
-
-    /// Build the volatile recalled-context block for a prompt: relevant
-    /// memories (similarity + domain overlap + recency via the context
-    /// store) and relevant skills (lexical + domain selection). `None` when
-    /// nothing relevant surfaced — an empty block would only burn cache.
-    ///
-    /// **Quarantine filter (Proposal 3):** frames whose memory id (`nod_…`)
-    /// appears in the session's `quarantined_ids` set are dropped before
-    /// rendering. These are memories cited untruthful ≥ 2 times — surfacing
-    /// them is active harm.
-    ///
-    /// **A/B control (Proposal 4):** when `ab_suppressed` is true (set by a
-    /// deterministic coin flip), recall returns `None` so the turn runs
-    /// without context — the outcome is then comparable to recalled turns.
-    pub async fn recall_block(&self, prompt: &str) -> Option<String> {
-        if self.ab_suppressed {
-            return None;
-        }
-        let mut sections: Vec<String> = Vec::new();
-        let frames = self.recalled_frames(prompt).await.frames;
-        if let Some(section) = render_context_section(&frames) {
-            sections.push(section);
-        }
-
-        let all_skills = self.load_skills();
-        let selected = skills::select_skills(
-            &all_skills,
-            prompt,
-            &self.domains.names(),
-            &SelectionConfig::default(),
-        );
-        if !selected.is_empty() {
-            sections.push(skills::render_skills_section(&selected));
-        }
-
-        // In-progress workspace maps belong here, not in the cached system
-        // prefix: a draft line names the producing pid and its liveness, so
-        // it differs per process and flips mid-session. Rendering it in the
-        // prefix is what made the prefix non-byte-stable (#639).
-        if let Some(section) = stella_tools::exploration::render_draft_claims(&self.workspace_root)
-        {
-            sections.push(section);
-        }
-
-        if sections.is_empty() {
-            None
-        } else {
-            Some(format!("{RECALL_MARKER}\n\n{}", sections.join("\n\n")))
-        }
-    }
-
-    /// A/B recall control (Proposal 4): suppress recall for this turn on a
-    /// deterministic `1/rate` schedule, returning whether recall was
-    /// suppressed. A rate of 0 (or 1) never suppresses.
-    ///
-    /// The flag is turn-scoped but only *reset* by this call, so every driver
-    /// that recalls must call it once per turn or inherit the previous turn's
-    /// arm. Today only the interactive REPL's plain prompts do
-    /// (`agent::run_interactive`); the outcome is attributed by tagging the
-    /// episode summary `[ab-control]` (see `agent::record_turn_episode`), which
-    /// is the whole readout — there is no reporting command yet, so the
-    /// comparison is an offline query over the episode store.
-    ///
-    /// Suppression is driven by a per-session **turn counter**, not a wall
-    /// clock. A previous implementation seeded off `SystemTime` nanoseconds
-    /// and tested `ns % rate == 0`; on any host whose realtime clock is
-    /// coarser than nanoseconds (macOS keeps it in microseconds, so `ns` is
-    /// always a multiple of 1000) that predicate is true on *every* turn for
-    /// any `rate` dividing 1000 — silently disabling recall entirely. A plain
-    /// counter makes exactly every `rate`-th turn a control turn, on every OS.
-    pub fn maybe_suppress_recall(&mut self, rate: u32) -> bool {
-        if rate == 0 || rate == 1 {
-            self.ab_suppressed = false;
-            return false;
-        }
-        self.ab_turn = self.ab_turn.wrapping_add(1);
-        self.ab_suppressed = ab_control_turn(self.ab_turn, rate);
-        self.ab_suppressed
-    }
-
-    /// Whether recall was suppressed this turn (for outcome attribution).
-    pub fn recall_was_suppressed(&self) -> bool {
-        self.ab_suppressed
-    }
-}
-
-/// Is the `turn`-th turn (1-based) an A/B control turn at the given `rate`?
-/// Every `rate`-th turn is a control turn; `rate` of 0 or 1 never controls.
-/// Pure so the schedule is property-testable independent of the (heavy)
-/// [`SessionMemory`] it lives on.
-fn ab_control_turn(turn: u32, rate: u32) -> bool {
-    rate > 1 && turn.is_multiple_of(rate)
 }
 
 impl SessionMemory {
@@ -469,452 +389,6 @@ impl SessionMemory {
         };
         let _ = self.store.upsert(delta).await;
     }
-
-    /// Post-turn self-reflection: one cheap model call producing 0-3
-    /// durable lessons, stored as domain-tagged reflection memories AND
-    /// appended to the skill-mining log; recurring lessons auto-promote to
-    /// SKILL.md files. Best-effort throughout — a failed reflection must never
-    /// fail the turn it describes. Returns a [`ReflectionReport`] so the caller
-    /// can surface the outcome (a model-call error, or how many lessons landed)
-    /// in whichever output format it speaks; the report distinguishes a genuine
-    /// model-call failure from the common, correct "nothing worth recording."
-    ///
-    /// `succeeded` controls the reflection prompt template (Proposal 1):
-    /// a failed turn gets a failure-analysis prompt that asks the model to
-    /// identify the root cause — the highest-value learning signal in the
-    /// system. A succeeded turn gets the conventional "what worked?" prompt.
-    pub async fn reflect_and_record(
-        &mut self,
-        provider: &dyn Provider,
-        model_hint: &str,
-        transcript: &[CompletionMessage],
-        quiet: bool,
-        succeeded: bool,
-        budget_limit: Option<f64>,
-    ) -> ReflectionReport {
-        let lessons = match reflect_on_turn(
-            provider,
-            model_hint,
-            &self.workspace_root,
-            transcript,
-            &self.domains.names(),
-            succeeded,
-            budget_limit,
-        )
-        .await
-        {
-            Ok((lessons, cost_usd, events)) => (lessons, cost_usd, events),
-            // The single reflection model call errored. Report it up so the
-            // caller can warn (text) or emit an event (stream-json) — this
-            // is the fix for the previously-silent reflection failure. Never
-            // fatal: the turn already stands on its own.
-            Err(model_error) => {
-                return ReflectionReport {
-                    recorded: 0,
-                    model_error: Some(model_error.message),
-                    cost_usd: model_error.cost_usd,
-                    events: model_error.events,
-                };
-            }
-        };
-        let (lessons, reflection_cost_usd, reflection_events) = lessons;
-
-        // Drop anything the user has already forgotten, BEFORE it reaches any
-        // of the three places this function persists to. Matching is by
-        // restatement, not equality: the loop re-learns paraphrases, so a
-        // lesson mined today can be the same lesson the user deleted last
-        // week wearing slightly different words. Suppressing here rather than
-        // at recall is what makes forgetting durable — an unsuppressed lesson
-        // would land in the log and stay re-mineable forever.
-        let lessons = self.retain_unforgotten(lessons);
-
-        if lessons.is_empty() {
-            return ReflectionReport {
-                cost_usd: reflection_cost_usd,
-                events: reflection_events,
-                ..ReflectionReport::default()
-            };
-        }
-
-        // 1. Store as recallable, domain-tagged reflection memories. Still
-        // best-effort (a failed reflection never fails the turn), but the
-        // outcome is kept so the "remembered" line below can't claim success
-        // for lessons that never landed in the store.
-        let delta = ContextDelta {
-            memories: lessons
-                .iter()
-                .map(|l| MemoryInput::reflection(&l.lesson, l.domains.iter().cloned()))
-                .collect(),
-            ..Default::default()
-        };
-        let stored = self.store.upsert(delta).await.is_ok();
-
-        // 2. Append to the mining log and mine for auto-creatable skills.
-        // Count how many lessons actually reached the log so the message below
-        // reports partial persistence accurately (some serialize/append writes
-        // may fail while others succeed).
-        let log_path =
-            stella_store::workspace_private_state_path(&self.workspace_root, "reflections.jsonl")
-                .ok();
-        let mut logged_count = 0usize;
-        for lesson in &lessons {
-            if let Ok(line) = serde_json::to_string(lesson)
-                && stella_store::append_workspace_private_line(
-                    &self.workspace_root,
-                    "reflections.jsonl",
-                    &line,
-                )
-                .is_ok()
-            {
-                logged_count += 1;
-            }
-        }
-        if let Some(log_path) = &log_path {
-            self.auto_create_skills(log_path, quiet);
-        }
-
-        if !quiet {
-            let n = lessons.len();
-            if stored {
-                println!(
-                    "  {} remembered {n} lesson(s) from this turn",
-                    "✦".magenta()
-                );
-            } else if logged_count == n {
-                println!(
-                    "  {} could not persist {n} lesson(s) to the context store \
-                     (logged to reflections.jsonl only)",
-                    "!".yellow()
-                );
-            } else if logged_count > 0 {
-                println!(
-                    "  {} could not persist {n} lesson(s) to the context store; \
-                     {logged_count} of {n} reached reflections.jsonl",
-                    "!".yellow()
-                );
-            } else {
-                println!(
-                    "  {} could not persist {n} lesson(s) — both the context store \
-                     and reflections.jsonl writes failed",
-                    "!".yellow()
-                );
-            }
-        }
-        ReflectionReport {
-            recorded: if stored { lessons.len() } else { 0 },
-            model_error: None,
-            cost_usd: reflection_cost_usd,
-            events: reflection_events,
-        }
-    }
-
-    /// Drop lessons that restate something the user forgot.
-    ///
-    /// Best-effort in one direction only: if the tombstone table cannot be
-    /// read we keep every lesson rather than silently discarding the turn's
-    /// learning. That is the opposite posture from recall — there, an
-    /// unreadable suppression set means surface nothing; here it means
-    /// persist everything. Both choices fail toward the recoverable outcome:
-    /// a memory that should have been suppressed can be forgotten again,
-    /// whereas a lesson dropped on a transient read error is gone for good.
-    fn retain_unforgotten(&self, lessons: Vec<ReflectionLesson>) -> Vec<ReflectionLesson> {
-        let forgotten = match stella_store::Store::open(&self.workspace_root)
-            .and_then(|store| store.forgotten_texts(stella_store::ContextSurface::Memory))
-        {
-            Ok(texts) if !texts.is_empty() => texts,
-            _ => return lessons,
-        };
-        lessons
-            .into_iter()
-            .filter(|l| {
-                !stella_store::is_suppressed(&l.lesson, forgotten.iter().map(String::as_str))
-            })
-            .collect()
-    }
-
-    /// Mine the whole reflection log for recurring lessons and auto-create
-    /// skills for any that qualify (threshold + session cap + no-clobber
-    /// enforced by `stella_core::skills`).
-    ///
-    /// The log is append-only and re-read in full every reflection turn, so
-    /// lines written before a tombstone existed are still in it. Filtering
-    /// the observations here is what stops a forgotten lesson from returning
-    /// as an auto-created skill — the second door that made a plain `DELETE`
-    /// insufficient.
-    fn auto_create_skills(&mut self, log_path: &Path, quiet: bool) {
-        let Ok(log) = std::fs::read_to_string(log_path) else {
-            return;
-        };
-        // Every surface a background loop can regenerate; `ContextSurface` owns which.
-        let forgotten: Vec<String> = stella_store::Store::open(&self.workspace_root)
-            .and_then(|store| {
-                let mut texts = Vec::new();
-                for surface in stella_store::ContextSurface::restatement_suppressing() {
-                    texts.extend(store.forgotten_texts(surface)?);
-                }
-                Ok(texts)
-            })
-            .unwrap_or_default();
-
-        let observations: Vec<SkillObservation> = log
-            .lines()
-            .filter_map(|line| serde_json::from_str::<ReflectionLesson>(line).ok())
-            .filter(|l| {
-                !stella_store::is_suppressed(&l.lesson, forgotten.iter().map(String::as_str))
-            })
-            .map(|l| SkillObservation {
-                reference: format!("reflection:{}", l.occurred_at),
-                text: l.lesson,
-                domains: l.domains,
-                occurred_at: l.occurred_at,
-                salient: false,
-            })
-            .collect();
-        if observations.is_empty() {
-            return;
-        }
-
-        let existing = self.load_skills();
-        let candidates =
-            skills::mine_skill_candidates(observations, &existing, &SkillMineConfig::default());
-
-        let skills_dir = self.workspace_skills_dir();
-        let existing_paths: Vec<String> = existing.iter().map(|s| s.source_path.clone()).collect();
-        let config = AutoCreateConfig::default();
-        for candidate in candidates {
-            match skills::decide_auto_creation(
-                &candidate,
-                &skills_dir,
-                &existing_paths,
-                self.skills_created,
-                &config,
-            ) {
-                AutoCreateDecision::Create { path } => {
-                    let markdown = skills::render_skill_markdown(&candidate);
-                    let path = PathBuf::from(path);
-                    if let Some(parent) = path.parent() {
-                        let _ = std::fs::create_dir_all(parent);
-                    }
-                    if std::fs::write(&path, markdown).is_ok() {
-                        self.skills_created += 1;
-                        if !quiet {
-                            println!(
-                                "  {} new skill auto-created from recurring observations: {} ({})",
-                                "✦".magenta().bold(),
-                                candidate.name.bright_magenta(),
-                                path.display()
-                            );
-                        }
-                    }
-                }
-                AutoCreateDecision::Skip { .. } => {}
-            }
-        }
-    }
-
-    /// Authoritative prompt and pipeline recall, including a fresh quarantine
-    /// read so prior-turn feedback applies immediately.
-    async fn recalled_frames(&self, goal: &str) -> Recall {
-        self.recalled_frames_reporting(goal, |message| eprintln!("  {} {message}", "!".yellow()))
-            .await
-    }
-    /// Recall with an injectable diagnostic sink to avoid global stderr capture in tests.
-    async fn recalled_frames_reporting(
-        &self,
-        goal: &str,
-        mut report: impl FnMut(String),
-    ) -> Recall {
-        if self.ab_suppressed {
-            return Recall::default();
-        }
-        let query = ContextQuery {
-            goal: goal.to_string(),
-            query_text: Some(goal.to_string()),
-            embedding: None,
-            kinds: vec![],
-            // Workspace files the goal names verbatim, as anchors: the
-            // code-graph provider answers anchors with each file's graph
-            // NEIGHBORHOOD (symbols + imports + importers), not just
-            // goal-token definition lookups — deterministic localization
-            // into both the planner prompt and the worker's recall block,
-            // instead of hoping the model's first move is a graph_query
-            // (#342 seam 3).
-            anchors: goal_path_anchors(goal, &self.workspace_root),
-            // Settings-backed since #712 deliverable 8; the defaults are the
-            // literals that used to be here.
-            max_frames: self.retrieval.max_frames,
-            max_tokens: self.retrieval.max_tokens,
-            as_of: None,
-            representation_preferences: vec![],
-        };
-        // Two independent suppression sets, read together and applied as one.
-        // Quarantine is *derived* — the model kept calling a memory untruthful;
-        // a tombstone is *stored* — a person read it and said remove it. Both
-        // are fail-closed for the same reason: if the state cannot be read,
-        // surfacing everything is the one outcome that is definitely wrong.
-        let quarantined = match stella_store::Store::open(&self.workspace_root).and_then(|store| {
-            let mut ids = store.quarantined_memory_ids()?;
-            ids.extend(store.forgotten_ids(stella_store::ContextSurface::Memory)?);
-            Ok(ids)
-        }) {
-            Ok(ids) => ids,
-            Err(error) => {
-                report(format!(
-                    "memory recall disabled: suppression state unavailable: {error}"
-                ));
-                return Recall::default();
-            }
-        };
-        // The usage report accounts for the fan-out that produced this turn's
-        // context, so it is captured even when quarantine or the citation-label
-        // filter later drops every frame: a provider still spent the tokens,
-        // and a cost that vanishes because the host discarded the frames is
-        // exactly the unmeterable cost #452 exists to surface.
-        let recalled = crate::contextgraph::recall_via_host(&self.host, &query).await;
-        Recall {
-            frames: recalled
-                .frames
-                .into_iter()
-                .filter_map(project_recalled_frame)
-                .filter(|frame| !is_suppressed_local_frame(frame, &quarantined))
-                .collect(),
-            usage: Some(recalled.usage),
-        }
-    }
-}
-
-/// Workspace files the goal names verbatim (`src/driver.rs`,
-/// `stella-core/src/bus.rs`), for use as recall anchors (#342 seam 3).
-/// Deterministic and cheap: path-shaped tokens (must contain `/`) that
-/// resolve to real files under the workspace root. Escapes are rejected,
-/// duplicates collapse, and the list is capped — anchors fan out into
-/// whole graph neighborhoods, so a pathological goal must not turn recall
-/// into a full-graph walk.
-fn goal_path_anchors(goal: &str, root: &std::path::Path) -> Vec<String> {
-    const MAX_ANCHORS: usize = 4;
-    let mut seen = std::collections::HashSet::new();
-    let mut anchors = Vec::new();
-    'tokens: for token in goal.split(|c: char| c.is_whitespace() || "\"'`()[]{}<>,;!?".contains(c))
-    {
-        // Strip trailing sentence punctuation (`fix src/a.rs.`) without
-        // touching the extension dot; a `src/a.rs:42` file:line spelling
-        // anchors the file.
-        let token = token.trim_end_matches(['.', ':']);
-        for candidate in [token, token.split(':').next().unwrap_or(token)] {
-            if !candidate.contains('/')
-                || candidate.len() > 256
-                || candidate.split('/').any(|seg| seg == "..")
-                || candidate.starts_with('/')
-            {
-                continue;
-            }
-            if root.join(candidate).is_file() && seen.insert(candidate.to_string()) {
-                anchors.push(candidate.to_string());
-                if anchors.len() == MAX_ANCHORS {
-                    break 'tokens;
-                }
-                break;
-            }
-        }
-    }
-    anchors
-}
-
-/// The pipeline's context-recall port over the workspace memory store: the
-/// split-context planner (L-E6) receives the same durable lessons the
-/// worker's injected recall block carries, as structured frames instead of a
-/// rendered string. Frames without a citation label are dropped (L-C4), and
-/// failed recall, including quarantine verification, degrades to no frames (L-C6).
-#[async_trait::async_trait]
-impl ContextRecallPort for SessionMemory {
-    async fn recall(&self, goal: &str) -> Recall {
-        self.recalled_frames(goal).await
-    }
-}
-
-/// Render recalled frames as the "Relevant context" section of the recall
-/// block. Memory-kind frames carry their stable `[nod_…]` id inline — the
-/// handle the `cite_memory` tool ties feedback to — and their presence
-/// appends the citation instruction, so the model is asked to cite exactly
-/// when there is something citable. Other frame kinds (code-graph hits,
-/// episodes) keep the plain label form: they are grounding, not memories,
-/// and never enter the citation → promotion loop. `None` when no frame has
-/// a citation label (L-C4 filters the rest).
-fn render_context_section(frames: &[RecalledFrame]) -> Option<String> {
-    let mut lines: Vec<String> = Vec::new();
-    let mut citable = false;
-    for f in frames {
-        let label = &f.citation_label;
-        if f.kind == "memory" {
-            citable = true;
-            if let Some(id) = &f.id {
-                lines.push(format!("- [{id}] {label} — {}", f.content.trim()));
-            }
-        } else {
-            lines.push(format!("- {} — {}", label, f.content.trim()));
-        }
-    }
-    if lines.is_empty() {
-        return None;
-    }
-    let mut section = format!("Relevant context:\n{}", lines.join("\n"));
-    if citable {
-        section.push_str(
-            "\n\nWhen a memory above (a [nod_…]-tagged line) actually informs your work this \
-             turn, call cite_memory with that id once you can judge it: useful_score 1-5 for \
-             how much it helped the actual work, truthful for whether its content still holds \
-             (verify against the workspace, don't assume), and a one-sentence remark. Cite \
-             only memories you genuinely used — no courtesy citations.",
-        );
-    }
-    Some(section)
-}
-
-/// Land the recalled-context message for this turn at the conversation
-/// TAIL — just before the turn's prompt when the prompt is already present
-/// (one-shot paths), appended otherwise (interactive paths push the prompt
-/// right after) — leaving previous turns' blocks in place as durable
-/// history. Rewriting or removing an early message every turn (the old
-/// index-1 refresh) byte-changed the front of the replayed history, which
-/// reduced the provider cache's reusable prefix to the system message
-/// alone for the whole session — the exact full-rate re-bill L-E8 exists
-/// to prevent. Durability's cost is bounded: an unchanged block is not
-/// re-appended (the model already sees it), so only genuinely new recall
-/// content adds tokens, and it rides the cached prefix from the next turn
-/// on. `None` (nothing relevant, or an A/B-suppressed turn) adds nothing
-/// and touches nothing.
-pub fn inject_recall_block(messages: &mut Vec<CompletionMessage>, block: Option<String>) {
-    let is_marker =
-        |m: &CompletionMessage| m.role == MessageRole::User && m.content.starts_with(RECALL_MARKER);
-    let Some(content) = block else { return };
-    if messages
-        .iter()
-        .rev()
-        .find(|m| is_marker(m))
-        .is_some_and(|m| m.content == content)
-    {
-        return;
-    }
-    let message = CompletionMessage {
-        role: MessageRole::User,
-        content,
-        tool_calls: vec![],
-        tool_results: vec![],
-        attachments: Vec::new(),
-    };
-    // Context precedes the question: when the turn's prompt is already the
-    // final message, slot the block just before it.
-    let at = match messages.last() {
-        Some(last)
-            if last.role == MessageRole::User
-                && !is_marker(last)
-                && last.tool_results.is_empty() =>
-        {
-            messages.len() - 1
-        }
-        _ => messages.len(),
-    };
-    messages.insert(at, message);
 }
 
 /// Seconds since the Unix epoch — the episode timestamps' primitive.
